requirements.txt
# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database (async)
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1

# Auth & Security
passlib[bcrypt]==1.7.4
PyJWT==2.8.0

# Redis (async)
redis[hiredis]==5.0.4

# Celery (task queue)
celery==5.4.0
flower==2.0.1

# Configuration
pydantic-settings==2.2.1
python-dotenv==1.0.1

# QR Code
qrcode[pil]==7.4.2
Pillow==10.3.0

# User-Agent parsing
user-agents==2.2.0

# HTTP client (for geo-lookup)
httpx==0.27.0

# URL validation
validators==0.28.3

# Testing
pytest==8.2.0
pytest-asyncio==0.23.7
httpx==0.27.0
anyio==4.3.0
aiosqlite==0.20.0

# Type stubs
types-redis==4.6.0.20240417
